Join prev and next on a common function
